// src/api/generated/model/eventTaskSubmissionDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */
import type { EventTaskSubmissionDtoStatus } from './eventTaskSubmissionDtoStatus';
import type { EventTaskSubmissionDtoReward } from './eventTaskSubmissionDtoReward';
import type { EventSubmissionItemDto } from './eventSubmissionItemDto';
import type { EventSubmissionEventDto } from './eventSubmissionEventDto';

export interface EventTaskSubmissionDto {
  id: string;
  taskId: string;
  ambassadorId: string;
  createdAt: string;
  updatedAt: string;
  status: EventTaskSubmissionDtoStatus;
  /** @nullable */
  comment: string | null;
  /** @nullable */
  reviewComment: string | null;
  /** Опыт только за окончательно одобренный ответ */
  experiencePoints: number;
  /** @nullable */
  reward: EventTaskSubmissionDtoReward;
  items: EventSubmissionItemDto[];
  events: EventSubmissionEventDto[];
}